Add Parse trait, and implement on Timestamp types.
Add a new public trait, Parse, which provides a method to parse an
element directly from a string. We can't use FromStr, TryFrom, etc,
because many elements retain a lifetime dependence on the string.

Add a blanket implementation of Parse for any type which implements
TryFrom<&str> for elements without a lifetime.