Fix drawer parsing to not take headlines.
Currently, when parsing a drawer for a node of level N, any headlines of
level N + 1 will be absorbed into the drawer. This changes drawer
parsing to reject headlines. For example:

```
* Hello
:PROPERTIES:
** World
:END:
```

should be parsed as two nodes. The drawer is ill-formed, and will be
parsed as part of the body instead of as a drawer.